Related UART code generation and clean the app.c
In the `app.c`, I change the AppTask1, 2, 3 series to **TestTask**. This task check the each sensors. So, if you check all the sensors then you have to remove that task.

And I make the code related in UART. When I add this code. I make the directory and add that directory to project build options.